api/requirements.txt
Flask==2.0.3
pexels_api_py==0.0.5
python-dotenv==0.20.0
shotstack_sdk==0.2.5
boto3
pexels_api_py==0.0.5
python-dotenv==0.20.0
shotstack_sdk==0.2.5
spacy
pytextrank